Refactor : 겹치는 부분 함수화
